backend/requirements.txt
# Core Framework
fastapi==0.117.1
uvicorn==0.37.0
sqlmodel==0.0.25
python-dotenv==1.1.1

# AI & Scraping
openai==2.0.0
requests==2.32.5
beautifulsoup4==4.12.3